remove unnecessary casts and add assertion
also clean test-case

+++ runtime/src/backends/basic/expr/mod.rs
@@ -235,7 +235,6 @@ impl ConstValue {
                 is_signed,
             }: IntType, // pattern matching in function args, cool!
         ) -> Option<u128> {
-            // TODO: please double check that my bit twiddling never accidentally overflows
             // we don't want any wrapping in this function, so we take the 0th parameter
             let (first, second) = (first.0, second.0);
             if is_signed {
@@ -252,9 +251,9 @@ impl ConstValue {
                 if let Some(result) = result {
                     // case: i128 has not overflowed, so result is valid. Check if we're
                     // higher than our type's max value or lower than it's min value.
-                    let max = ((1_u128 << (bit_size as u128 - 1)) as u128 - 1) as i128;
+                    let max = ((1_u128 << (bit_size - 1)) - 1) as i128;
                     let min = match bit_size {
-                        0..=127 => -(1_i128 << (bit_size as i128 - 1)),
+                        0..=127 => -(1_i128 << (bit_size - 1)),
                         128 => i128::MIN,
                         129..=u64::MAX => panic!("unsupported integer size; too large"),
                     };
@@ -278,7 +277,7 @@ impl ConstValue {
                 if let Some(result) = result {
                     // case: u128 has not overflowed, check if we're higher than our max
                     let max = match bit_size {
-                        0..=127 => (1_u128 << (bit_size as u128)) - 1,
+                        0..=127 => (1_u128 << bit_size) - 1,
                         128 => u128::MAX,
                         129..=u64::MAX => panic!("unsupported integer size; too large"),
                     };
